Parse JSON using json library
Previously, the code was reading data from a URL provided by the user,
then executing the data as if it were Python source code using the eval
function. This is not safe. There is no guarantee that the remote Jenkins
is not controlled by my evil twin brother from a parallel universe.

This changes the code to parse the data, which should be JSON, using
the Python standard library's json module. I've not actually tested
this (there is no test suite), but it should do the same thing.