get_downstream_builds doesn't return the proper data
build.get_downstream_builds doesn't return all of the downstream jobs when a job calls multiple of the same job.  Jenkins shows this information by giving you the range of build numbers.  This change fixes that by returning you all of the downstream builds within that range instead of just the first build.