Add plugin uninstall; install exact versions
of plugins.

Previously, installing plugins would result in a POST to
pluginManager/installNecessaryPlugins, but this endpoint will always
ignore the requested version and install the latest version. Plugin
versions that are not the latest version can only be installed by
uploading the hpi/jpi file for that version.

To get the previous versions' hpi file, the update-center.json file will
be parsed to retrive the download URL for the current version.  The
current version number in the url will then be replaced by the requested
version number. The plugin will be downloaded, unpacked and eventual
dependencies (at their latest version) will be installed.

This commit also introduces the ability to uninstall plugins and adds an
option to wait for reboot (moved from test_restart integration test)
when calling safe_restart().