API: add toggle keep build forever
This commit aims to add an API function that allows a user of the
API to toggle the "keep this build forever" flag on the jenkins
build page.

This is related to #645 but does not close it as it's missing the
"edit the build information - display name and description"
part of the summary.

https://github.com/pycontribs/jenkinsapi/issues/645